debugvmfs cat: Support files bigger than 2GB

+++ vmfs_file.c
@@ -252,7 +252,7 @@ ssize_t vmfs_file_read(vmfs_file_t *f,u_char *buf,size_t len)
 }
 
 /* Dump a file */
-int vmfs_file_dump(vmfs_file_t *f,off_t pos,size_t len,FILE *fd_out)
+int vmfs_file_dump(vmfs_file_t *f,off_t pos,uint64_t len,FILE *fd_out)
 {
    u_char *buf;
    ssize_t res;
